+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # encapsule releases
 
 ## 0.5 (2026-08-15)
+- `commit`: add `--name` to set the encapsule image name
 - `commit`: replaces `refresh` with simpler logic
 - `create`,`run`: now always act on a container image
 - bind-mount `/etc/localtime`